Fixed support for Type 2 fonts
Type 2 fonts are now scanned correctly but I had to comment a line in
CMap as it would eat up the whole description of the CMap. It might
need to be changed as I'm not sure if that could lead to other kind of
problems.
Right now the only problem left is that the highlighting of the
characters for Type 2 fonts is not correct.